fix(component): fix blocking delays and multiple http calls (#45)
Updating the logic for starting / stopping the blocker so that the state of blocking is saved, then whenever a start, stop or reset event is received, it checks whether other blocked calls are already in progress and processes the call accordingly.
This should allow multiple concurrent calls to be handled, only showing the blocker at the start of the first call and not hiding it until the end of the final call.
Also, if a delayStart and delayStop value is set, the delay will only be used for hiding the blocker if the blocker is already active when the stop event is raised rather than processing the entire stop event after the delay, regardless of the current state of the blocker. This should stop the issue where the blocker can flicker on and off after a short http call, no matter what.